Replaced weighted mean with if statements
I hate weighted means, so I got rid of it. Using if statements to give
outside sensors priority over inside sensors when calculating position.

+++ MAE508_finalcode/MAE508_finalcode.ino
@@ -56,7 +56,7 @@ unsigned int sensorValues[NUM_SENSORS];
 int averagePosition(int *sampleBuffer)
 {
   int i;
-  double accum = 0;
+  long accum = 0;
   int average = 0;
   for(i = 0; i < SAMPLE_BUFFER_SIZE; i++)
   {
